1330 henter om bruker har søknad til behandling hvis de ikke har andre meldekort
