+++ src/main/kotlin/no/nav/tiltakspenger/meldekort/service/BrukerService.kt
@@ -1,6 +1,9 @@
 package no.nav.tiltakspenger.meldekort.service
 
+import arrow.core.getOrElse
 import no.nav.tiltakspenger.libs.common.Fnr
+import no.nav.tiltakspenger.meldekort.clients.saksbehandling.SaksbehandlingClient
+import no.nav.tiltakspenger.meldekort.domene.ArenaMeldekortStatus
 import no.nav.tiltakspenger.meldekort.domene.Bruker
 import no.nav.tiltakspenger.meldekort.domene.Sak
 import no.nav.tiltakspenger.meldekort.repository.SakRepo
@@ -9,6 +12,7 @@ class BrukerService(
     private val meldekortService: MeldekortService,
     private val sakRepo: SakRepo,
     private val arenaMeldekortStatusService: ArenaMeldekortStatusService,
+    private val saksbehandlingClient: SaksbehandlingClient,
 ) {
     suspend fun hentBruker(fnr: Fnr): Bruker {
         val sak = sakRepo.hentTilBruker(fnr)
@@ -33,10 +37,17 @@ class BrukerService(
 
     private suspend fun hentBrukerUtenSak(fnr: Fnr): Bruker.UtenSak {
         val arenaMeldekortStatus = arenaMeldekortStatusService.hentArenaMeldekortStatus(fnr)
+        val harSoknadUnderBehandling = if (arenaMeldekortStatus != ArenaMeldekortStatus.HAR_MELDEKORT) {
+            // Defaulter til false hvis kallet til saksbehandling-api feiler siden det er viktigere at de får opp siden enn at feilmeldingen er korrekt
+            saksbehandlingClient.harSoknadUnderBehandling(fnr).getOrElse { false }
+        } else {
+            false
+        }
 
         return Bruker.UtenSak(
             fnr = fnr,
             arenaMeldekortStatus = arenaMeldekortStatus,
+            harSoknadUnderBehandling = harSoknadUnderBehandling,
         )
     }
 }
